Force symlink permissions to 777 in layers
This is the permissions set on Linux, and this is what we should use.
Without this on Darwin we can get layer digest mismatch, because on
Darwin symlinks in /nix/store have 755 permissions instead. Nix
generally ignores permissions in /nix/store paths.
See https://github.com/nlewo/nix2container/issues/23 for details.

Fixes: #23