Fixes #94: exclude config file and json files from nix database.
`closureGraph` is extended to ignore all paths passed in the new
ignore arguement.

The resulting graph is both passed to nix2container-bin and
makeNixDatabase -- ensuring that both contain the same paths.

makeNixDatabase is rewritten, such that it always expects a
closureGraphJson as arguement but does not include it in the database.

The `--ignore` flag of the nix2container-bin is not used anymore.

https://github.com/nlewo/nix2container/issues/94